Login

+++ app/src/main/java/com/example/plantaid_redesign/Model/Main.java
@@ -5,7 +5,7 @@ public class Main {
     private double feels_like;
     private double temp_min;
     private double temp_max;
-    private int pressure;
+    private float pressure;
     private int humidity;
     private int sea_level;
     private int grnd_level;
@@ -45,7 +45,7 @@ public class Main {
         this.temp_max = temp_max;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
